requirements.txt
numpy>=1.21
torch>=2.1.0
torchvision>=0.16.0
pillow>=9.0.0
tqdm>=4.62.3
SoccerNet>=0.1.23
opencv-python>=4.8.1.78
matplotlib>=3.5.1